backend/requirements.txt
fastapi>=0.115.12,<0.116.0
uvicorn[standard]>=0.34.3,<0.35.0
sqlalchemy[asyncio]>=2.0.41,<3.0.0
asyncpg>=0.30.0,<0.31.0
alembic>=1.16.1,<2.0.0
pydantic[dotenv]>=2.11.5,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
passlib[argon2]>=1.7.4,<2.0.0
cryptography>=45.0.3,<46.0.0
python-multipart>=0.0.9,<0.1.0
email-validator>=2.1.0,<3.0.0
aiosqlite>=0.21.0,<0.22.0 